Refocus explanation of clone-pointers question
I find it clearer to explain this solution in terms of what type
parameter the generic function is instantiated with.

To facilitate that, I renamed the type parameter from T to X so that we
don't mix it up with talking about references &T and Rc<T> where the
symbol T does not refer specifically to the type parameter of our
specific function in the quiz code.

Filled in some relevant detail as well:

- Why does impl Clone for &T exist if it is so confusing?

- Why does b.clone() not call the impl for &()?