Fix up the explanation of question 37.
Before this patch the explanation to this question wasn't completely correct,
as the creation of temporaries isn't linked to the use of wildcard expressions.

This is important because following the current explanation something like

```rs
let _ = Drop0;
println!("1");

```

Would print `10` while the opposite is the case, since lifetime extension based on
expressions only happens for the operands of borrow operators.

This patch updates the explanation to be a bit more precise according to the reference
and it also corrects a place where there was a mix up between wildcard pattern and
underscore expression.

Fixes #91